requirements.txt
discord.py==2.5.2
aiohttp==3.12.15
aiosqlite==0.21.0
python-dotenv==1.1.1
